TECS: Replaced old tecs by cleaned up version.
