+++ src/lib/tecs/TECS.cpp
@@ -30,6 +30,11 @@
  * POSSIBILITY OF SUCH DAMAGE.
  *
  ****************************************************************************/
+/**
+ * @file TECS.cpp
+ *
+ * @author Paul Riseborough
+ */
 
 #include "TECS.hpp"
 
@@ -41,242 +46,423 @@ using math::constrain;
 using math::max;
 using math::min;
 
-static constexpr float DT_MIN = 0.001f;	///< minimum allowed value of _dt (sec)
-static constexpr float DT_MAX = 1.0f;	///< max value of _dt allowed before a filter state reset is performed (sec)
-
-/**
- * @file TECS.cpp
- *
- * @author Paul Riseborough
- */
-
-/*
- * This function implements a complementary filter to estimate the climb rate when
- * inertial nav data is not available. It also calculates a true airspeed derivative
- * which is used by the airspeed complimentary filter.
- */
-void TECS::update_vehicle_state_estimates(float equivalent_airspeed, const float speed_deriv_forward,
-		bool altitude_lock, float altitude, float vz)
+// TODO there seems to be an inconsistent definition of IAS/CAS/EAS/TAS
+// TODO Recheck the timing.
+void TECSAirspeedFilter::initialize(const float equivalent_airspeed)
 {
-	// calculate the time lapsed since the last update
-	uint64_t now = hrt_absolute_time();
-	float dt = fmaxf((now - _state_update_timestamp) * 1e-6f, DT_MIN);
 
-	bool reset_altitude = false;
 
-	if (_state_update_timestamp == 0 || dt > DT_MAX) {
-		dt = DT_DEFAULT;
-		reset_altitude = true;
-	}
-
-	if (!altitude_lock) {
-		reset_altitude = true;
-	}
-
-	if (reset_altitude) {
-		_states_initialized = false;
-	}
-
-	_state_update_timestamp = now;
-	_EAS = equivalent_airspeed;
-
-	// Set the velocity and position state to the the INS data
-	_vert_vel_state = -vz;
-	_vert_pos_state = altitude;
-
-	// Update and average speed rate of change if airspeed is being measured
-	if (PX4_ISFINITE(equivalent_airspeed) && airspeed_sensor_enabled()) {
-		_tas_rate_raw = speed_deriv_forward;
-		// Apply some noise filtering
-		_TAS_rate_filter.update(speed_deriv_forward);
-		_tas_rate_filtered = _TAS_rate_filter.getState();
-
-	} else {
-		_tas_rate_raw = 0.0f;
-		_tas_rate_filtered = 0.0f;
-	}
+	_airspeed_state.speed = equivalent_airspeed;
+	_airspeed_state.speed_rate = 0.0f;
+	_airspeed_rate_filter.reset(0.0f);
 }
 
-void TECS::_update_speed_states(float equivalent_airspeed_setpoint, float equivalent_airspeed, float EAS2TAS)
+void TECSAirspeedFilter::update(const float dt, const Input &input, const Param &param,
+				const bool airspeed_sensor_available)
 {
-	// Calculate the time in seconds since the last update and use the default time step value if out of bounds
-	uint64_t now = hrt_absolute_time();
-	const float dt = constrain((now - _speed_update_timestamp) * 1.0e-6f, DT_MIN, DT_MAX);
+	// Input checking
+	if (!(PX4_ISFINITE(dt) && dt > FLT_EPSILON)) {
+		// Do not update the states.
+		PX4_WARN("Time intervall is not valid.");
+		return;
+	}
 
-	// Convert equivalent airspeed quantities to true airspeed
-	_EAS_setpoint = PX4_ISFINITE(equivalent_airspeed_setpoint) ? equivalent_airspeed_setpoint : _equivalent_airspeed_trim;
-	_TAS_setpoint  = _EAS_setpoint * EAS2TAS;
-	_TAS_max   = _equivalent_airspeed_max * EAS2TAS;
-	_TAS_min   = _equivalent_airspeed_min * EAS2TAS;
+	float airspeed;
 
-	// If airspeed measurements are not being used, fix the airspeed estimate to the nominal trim airspeed
-	if (!PX4_ISFINITE(equivalent_airspeed) || !airspeed_sensor_enabled()) {
-		_EAS = _equivalent_airspeed_trim;
+	if (PX4_ISFINITE(input.equivalent_airspeed) && airspeed_sensor_available) {
+		airspeed = input.equivalent_airspeed;
 
 	} else {
-		_EAS = equivalent_airspeed;
+		airspeed = param.equivalent_airspeed_trim;
 	}
 
-	// If first time through reset airspeed states
-	if (_speed_update_timestamp == 0) {
-		_tas_rate_state = 0.0f;
-		_tas_state = (_EAS * EAS2TAS);
+	float airspeed_derivative;
+
+	if (PX4_ISFINITE(input.equivalent_airspeed_rate) && airspeed_sensor_available) {
+		airspeed_derivative = input.equivalent_airspeed_rate;
+
+	} else {
+		airspeed_derivative = 0.0f;
 	}
 
-	// Obtain a smoothed TAS estimate using a second order complementary filter
+	// TODO remove. Only here for compatibility check with old TECS.
+	// filter true airspeed rate using first order filter with 0.5 second time constant
+	_airspeed_rate_filter.setParameters(TECS::DT_DEFAULT, param.speed_derivative_time_const);
+	_airspeed_rate_filter.reset(0.0f);
 
+	// Alpha filtering done in the TECS module. TODO merge with the second order complementary filter.
+	//_airspeed_rate_filter.setParameters(dt, param.speed_derivative_time_const);
+	if (PX4_ISFINITE(input.equivalent_airspeed_rate) && airspeed_sensor_available) {
+		_airspeed_rate_filter.update(airspeed_derivative);
+
+	} else {
+		_airspeed_rate_filter.reset(0.0f);
+	}
+
+	AirspeedFilterState new_airspeed_state;
 	// Update TAS rate state
-	_tas_innov = (_EAS * EAS2TAS) - _tas_state;
-	float tas_rate_state_input = _tas_innov * _tas_estimate_freq * _tas_estimate_freq;
+	float airspeed_innovation = airspeed - _airspeed_state.speed;
+	float airspeed_rate_state_input = airspeed_innovation * param.airspeed_estimate_freq * param.airspeed_estimate_freq;
+	new_airspeed_state.speed_rate = _airspeed_state.speed_rate + airspeed_rate_state_input * dt;
 
-	// Update TAS state
-	_tas_rate_state = _tas_rate_state + tas_rate_state_input * dt;
-	float tas_state_input = _tas_rate_state + _tas_rate_raw + _tas_innov * _tas_estimate_freq * 1.4142f;
-	const float new_tas_state = _tas_state + tas_state_input * dt;
+	// Update TAS state // TODO the airspeed rate is applied twice.
+	float airspeed_state_input = _airspeed_state.speed_rate + airspeed_derivative + airspeed_innovation *
+				     param.airspeed_estimate_freq * 1.4142f;
+	new_airspeed_state.speed = _airspeed_state.speed + airspeed_state_input * dt;
 
-	if (new_tas_state < 0.0f) {
-		// clip TAS at zero, back calculate rate
-		tas_state_input = -_tas_state / dt;
-		_tas_rate_state = tas_state_input - _tas_rate_raw - _tas_innov * _tas_estimate_freq * 1.4142f;
-		_tas_state = 0.0f;
-
-	} else {
-		_tas_state = new_tas_state;
+	// Clip tas at zero
+	if (new_airspeed_state.speed < 0.0f) {
+		// clip TAS at zero, back calculate rate // TODO Redo
+		airspeed_state_input = -_airspeed_state.speed / dt;
+		new_airspeed_state.speed_rate = airspeed_state_input - airspeed_derivative - airspeed_innovation *
+						param.airspeed_estimate_freq * 1.4142f;
+		new_airspeed_state.speed = 0.0f;
 	}
 
-	_speed_update_timestamp = now;
-
+	// Update states
+	_airspeed_state = new_airspeed_state;
 }
 
-void TECS::_update_speed_setpoint()
+TECSAirspeedFilter::AirspeedFilterState TECSAirspeedFilter::getState() const
 {
-	// Set the TAS demand to the minimum value if an underspeed or
-	// or a uncontrolled descent condition exists to maximise climb rate
-	if (_uncommanded_descent_recovery) {
-		_TAS_setpoint = _TAS_min;
+	AirspeedFilterState filter_state{
+		.speed = _airspeed_state.speed,
+		.speed_rate = _airspeed_rate_filter.getState()
+	};
 
-	} else if (_percent_undersped > FLT_EPSILON) {
-		// TAS setpoint is reset from external setpoint every time tecs is called, so the interpolation is still
-		// between current setpoint and mininimum airspeed here (it's not feeding the newly adjusted setpoint
-		// from this line back into this method each time).
-		// TODO: WOULD BE GOOD to "functionalize" this library a bit and remove many of these internal states to
-		// avoid the fear of side effects in simple operations like this.
-		_TAS_setpoint = _TAS_min * _percent_undersped + (1.0f - _percent_undersped) * _TAS_setpoint;
+	return filter_state;
+}
+
+void TECSReferenceModel::update(const float dt, const AltitudeReferenceState &setpoint, float altitude,
+				const Param &param)
+{
+	// Input checks
+	if (!(PX4_ISFINITE(dt) && dt > FLT_EPSILON)) {
+		// Do not update the states.
+		PX4_WARN("Time intervall is not valid.");
+		return;
 	}
 
-	_TAS_setpoint = constrain(_TAS_setpoint, _TAS_min, _TAS_max);
+	if (!PX4_ISFINITE(altitude)) {
+		altitude = 0.0f;
+	}
 
-	// Calculate limits for the demanded rate of change of speed based on physical performance limits
-	// with a 50% margin to allow the total energy controller to correct for errors.
-	const float max_tas_rate_sp = 0.5f * _STE_rate_max / math::max(_tas_state, FLT_EPSILON);
-	const float min_tas_rate_sp = 0.5f * _STE_rate_min / math::max(_tas_state, FLT_EPSILON);
+	// TODO rearrange handling of altitude rate and altitude. alt_rate should rather be a feedforward term.
+	float virtual_altitude_setpoint{setpoint.alt};
 
-	_TAS_setpoint_adj = constrain(_TAS_setpoint, _TAS_min, _TAS_max);
+	// Velocity setpoint reference
+	const bool input_is_altitude_rate = PX4_ISFINITE(setpoint.alt_rate);
+
+	_velocity_control_traj_generator.setMaxJerk(param.jerk_max);
+	_velocity_control_traj_generator.setMaxAccelUp(param.vert_accel_limit);
+	_velocity_control_traj_generator.setMaxAccelDown(param.vert_accel_limit);
+	_velocity_control_traj_generator.setMaxVelUp(param.max_sink_rate);
+	_velocity_control_traj_generator.setMaxVelDown(param.max_climb_rate);
+
+	if (input_is_altitude_rate) {
+		_velocity_control_traj_generator.setVelSpFeedback(setpoint.alt_rate);
+		_velocity_control_traj_generator.setCurrentPositionEstimate(altitude);
+		_velocity_control_traj_generator.update(dt, setpoint.alt_rate);
+		virtual_altitude_setpoint = _velocity_control_traj_generator.getCurrentPosition();
+
+	} else {
+		_velocity_control_traj_generator.reset(0.0f, 0.0f, altitude);
+	}
+
+	// Altitude setpoint reference
+	bool altitude_control_enable{PX4_ISFINITE(virtual_altitude_setpoint)};
+	_alt_control_traj_generator.setMaxJerk(param.jerk_max);
+	_alt_control_traj_generator.setMaxAccel(param.vert_accel_limit);
+	_alt_control_traj_generator.setMaxVel(fmax(param.max_climb_rate, param.max_sink_rate));
+
+	if (altitude_control_enable) {
+		float target_climbrate = math::min(param.target_climbrate, param.max_climb_rate);
+		float target_sinkrate = math::min(param.target_sinkrate, param.max_sink_rate);
+
+		const float delta_trajectory_to_target_m = setpoint.alt - _alt_control_traj_generator.getCurrentPosition();
+
+		float altitude_rate_target = math::signNoZero<float>(delta_trajectory_to_target_m) *
+					     math::trajectory::computeMaxSpeedFromDistance(
+						     param.jerk_max, param.vert_accel_limit, fabsf(delta_trajectory_to_target_m), 0.0f);
+
+		altitude_rate_target = math::constrain(altitude_rate_target, -target_sinkrate, target_climbrate);
+
+		_alt_control_traj_generator.updateDurations(altitude_rate_target);
+		_alt_control_traj_generator.updateTraj(dt);
+
+	} else {
+		_alt_control_traj_generator.reset(0.0f, 0.0f, altitude);
+	}
+}
+
+TECSReferenceModel::AltitudeReferenceState TECSReferenceModel::getAltitudeReference() const
+{
+	TECSReferenceModel::AltitudeReferenceState ref{
+		.alt = _alt_control_traj_generator.getCurrentPosition(),
+		.alt_rate = _alt_control_traj_generator.getCurrentVelocity(),
+	};
+
+	return ref;
+}
+
+float TECSReferenceModel::getAltitudeRateReference() const
+{
+	return _velocity_control_traj_generator.getCurrentVelocity();
+}
+
+void TECSReferenceModel::initialize(const AltitudeReferenceState &state)
+{
+	AltitudeReferenceState init_state{state};
+
+	if (!PX4_ISFINITE(state.alt)) {
+		init_state.alt = 0.0f;
+	}
+
+	if (!PX4_ISFINITE(state.alt_rate)) {
+		init_state.alt_rate = 0.0f;
+	}
+
+	_alt_control_traj_generator.reset(0.0f, init_state.alt_rate, init_state.alt);
+	_velocity_control_traj_generator.reset(0.0f, init_state.alt_rate, init_state.alt);
+}
+
+void TECSControl::initialize()
+{
+	_ste_rate_error_filter.reset(0.0f);
+	resetIntegrals();
+}
+
+void TECSControl::update(const float dt, const Setpoint &setpoint, const Input &input, Param &param, const Flag &flag)
+{
+	// Input checking
+	if (!(PX4_ISFINITE(dt) && dt > FLT_EPSILON)) {
+		// Do not update the states and output.
+		PX4_WARN("Time intervall is not valid.");
+		return;
+	}
+
+	AltitudePitchControl control_setpoint;
+
+	control_setpoint.tas_rate_setpoint = _airspeedControl(setpoint, input, param, flag);
+
+	control_setpoint.altitude_rate_setpoint = _altitudeControl(setpoint, input, param);
+
+	SpecificEnergy se{_updateEnergyBalance(control_setpoint, input)};
+
+	_detectUnderspeed(input, param, flag);
+
+	_updatePitchSetpoint(dt, input, se, param, flag);
+
+	_updateThrottleSetpoint(dt, se, param, flag);
+
+	_debug_output.altitude_rate_control = control_setpoint.altitude_rate_setpoint;
+	_debug_output.true_airspeed_derivative_control = control_setpoint.tas_rate_setpoint;
+}
+
+TECSControl::STELimit TECSControl::_calculateTotalEnergyRateLimit(const Param &param) const
+{
+	TECSControl::STELimit limit;
+	// Calculate the specific total energy rate limits from the max throttle limits
+	limit.STE_rate_max = math::max(param.max_climb_rate, FLT_EPSILON) * CONSTANTS_ONE_G;
+	limit.STE_rate_min = - math::max(param.max_sink_rate, FLT_EPSILON) * CONSTANTS_ONE_G;
+
+	return limit;
+}
+
+float TECSControl::_airspeedControl(const Setpoint &setpoint, const Input &input, const Param &param,
+				    const Flag &flag) const
+{
+	float airspeed_rate_output{0.0f};
+
+	STELimit limit{_calculateTotalEnergyRateLimit(param)};
 
 	// calculate the demanded true airspeed rate of change based on first order response of true airspeed error
 	// if airspeed measurement is not enabled then always set the rate setpoint to zero in order to avoid constant rate setpoints
-	if (airspeed_sensor_enabled()) {
-		_TAS_rate_setpoint = constrain((_TAS_setpoint_adj - _tas_state) * _airspeed_error_gain, min_tas_rate_sp,
-					       max_tas_rate_sp);
-
-	} else {
-		_TAS_rate_setpoint = 0.0f;
+	if (flag.airspeed_enabled) {
+		// Calculate limits for the demanded rate of change of speed based on physical performance limits
+		// with a 50% margin to allow the total energy controller to correct for errors.
+		const float max_tas_rate_sp = 0.5f * limit.STE_rate_max / math::max(input.tas, FLT_EPSILON);
+		const float min_tas_rate_sp = 0.5f * limit.STE_rate_min / math::max(input.tas, FLT_EPSILON);
+		airspeed_rate_output = constrain((setpoint.tas_setpoint - input.tas) * param.airspeed_error_gain, min_tas_rate_sp,
+						 max_tas_rate_sp);
 	}
 
+	return airspeed_rate_output;
 }
 
-void TECS::runAltitudeControllerSmoothVelocity(float alt_sp_amsl_m, float target_climbrate_m_s,
-		float target_sinkrate_m_s,
-		float alt_amsl)
+float TECSControl::_altitudeControl(const Setpoint &setpoint, const Input &input, const Param &param) const
 {
-	target_climbrate_m_s = math::min(target_climbrate_m_s, _max_climb_rate);
-	target_sinkrate_m_s = math::min(target_sinkrate_m_s, _max_sink_rate);
+	float altitude_rate_output;
+	altitude_rate_output = (setpoint.altitude_reference.alt - input.altitude) * param.altitude_error_gain +
+			       param.altitude_setpoint_gain_ff * setpoint.altitude_reference.alt_rate;
+	altitude_rate_output = math::constrain(altitude_rate_output, -param.max_sink_rate, param.max_climb_rate);
 
-	const float delta_trajectory_to_target_m = alt_sp_amsl_m - _alt_control_traj_generator.getCurrentPosition();
-
-	float height_rate_target = math::signNoZero<float>(delta_trajectory_to_target_m) *
-				   math::trajectory::computeMaxSpeedFromDistance(
-					   _jerk_max, _vert_accel_limit, fabsf(delta_trajectory_to_target_m), 0.0f);
-
-	height_rate_target = math::constrain(height_rate_target, -target_sinkrate_m_s, target_climbrate_m_s);
-
-	_alt_control_traj_generator.updateDurations(height_rate_target);
-	_alt_control_traj_generator.updateTraj(_dt);
-
-	_hgt_setpoint = _alt_control_traj_generator.getCurrentPosition();
-	_hgt_rate_from_hgt_ref = _alt_control_traj_generator.getCurrentVelocity();
-	_hgt_rate_setpoint = (_hgt_setpoint - alt_amsl) * _height_error_gain + _height_setpoint_gain_ff *
-			     _alt_control_traj_generator.getCurrentVelocity();
-	_hgt_rate_setpoint = math::constrain(_hgt_rate_setpoint, -_max_sink_rate, _max_climb_rate);
+	return altitude_rate_output;
 }
 
-void TECS::_detect_underspeed()
+TECSControl::SpecificEnergy TECSControl::_updateEnergyBalance(const AltitudePitchControl &control_setpoint,
+		const Input &input) const
 {
-	if (!_detect_underspeed_enabled) {
+	SpecificEnergy se;
+	// Calculate specific energy rate demands in units of (m**2/sec**3)
+	se.spe.rate_setpoint = control_setpoint.altitude_rate_setpoint * CONSTANTS_ONE_G; // potential energy rate of change
+	se.ske.rate_setpoint = input.tas * control_setpoint.tas_rate_setpoint; // kinetic energy rate of change
+
+	// Calculate specific energy rates in units of (m**2/sec**3)
+	se.spe.rate = input.altitude_rate * CONSTANTS_ONE_G; // potential energy rate of change
+	se.ske.rate = input.tas * input.tas_rate;// kinetic energy rate of change
+
+	// Calculate energy rate error
+	se.spe.rate_error = se.spe.rate_setpoint - se.spe.rate;
+	se.ske.rate_error = se.ske.rate_setpoint - se.ske.rate;
+
+	return se;
+}
+
+void TECSControl::_detectUnderspeed(const Input &input, const Param &param, const Flag &flag)
+{
+	if (!flag.detect_underspeed_enabled) {
 		_percent_undersped = 0.0f;
 		return;
 	}
 
 	// this is the expected (something like standard) deviation from the airspeed setpoint that we allow the airspeed
 	// to vary in before ramping in underspeed mitigation
-	const float tas_error_bound = kTASErrorPercentage * _equivalent_airspeed_trim;
+	const float tas_error_bound = param.tas_error_percentage * param.equivalent_airspeed_trim;
 
 	// this is the soft boundary where underspeed mitigation is ramped in
 	// NOTE: it's currently the same as the error bound, but separated here to indicate these values do not in general
 	// need to be the same
-	const float tas_underspeed_soft_bound = kTASErrorPercentage * _equivalent_airspeed_trim;
+	const float tas_underspeed_soft_bound = param.tas_error_percentage * param.equivalent_airspeed_trim;
 
-	const float tas_fully_undersped = math::max(_TAS_min - tas_error_bound - tas_underspeed_soft_bound, 0.0f);
-	const float tas_starting_to_underspeed = math::max(_TAS_min - tas_error_bound, tas_fully_undersped);
+	const float tas_fully_undersped = math::max(param.tas_min - tas_error_bound - tas_underspeed_soft_bound, 0.0f);
+	const float tas_starting_to_underspeed = math::max(param.tas_min - tas_error_bound, tas_fully_undersped);
 
-	_percent_undersped = 1.0f - math::constrain((_tas_state - tas_fully_undersped) /
+	_percent_undersped = 1.0f - math::constrain((input.tas - tas_fully_undersped) /
 			     math::max(tas_starting_to_underspeed - tas_fully_undersped, FLT_EPSILON), 0.0f, 1.0f);
 }
 
-void TECS::_update_energy_estimates()
+TECSControl::SpecificEnergyWeighting TECSControl::_updateSpeedAltitudeWeights(const Param &param, const Flag &flag)
 {
-	// Calculate specific energy demands in units of (m**2/sec**2)
-	_SPE_setpoint = _hgt_setpoint * CONSTANTS_ONE_G; // potential energy
-	_SKE_setpoint = 0.5f * _TAS_setpoint_adj * _TAS_setpoint_adj; // kinetic energy
 
-	// Calculate total energy error
-	_STE_error = _SPE_setpoint - _SPE_estimate + _SKE_setpoint - _SKE_estimate;
+	SpecificEnergyWeighting weight;
+	// Calculate the weight applied to control of specific kinetic energy error
+	float pitch_speed_weight = constrain(param.pitch_speed_weight, 0.0f, 2.0f);
 
-	// Calculate the specific energy balance demand which specifies how the available total
-	// energy should be allocated to speed (kinetic energy) and height (potential energy)
-	// Calculate the specific energy balance error
-	_SEB_error = SEB_setpoint() - (_SPE_estimate * _SPE_weighting - _SKE_estimate * _SKE_weighting);
+	if (flag.climbout_mode_active && flag.airspeed_enabled) {
+		pitch_speed_weight = 2.0f;
 
-	// Calculate specific energy rate demands in units of (m**2/sec**3)
-	_SPE_rate_setpoint = _hgt_rate_setpoint * CONSTANTS_ONE_G; // potential energy rate of change
-	_SKE_rate_setpoint = _tas_state * _TAS_rate_setpoint; // kinetic energy rate of change
+	} else if (_percent_undersped > FLT_EPSILON && flag.airspeed_enabled) {
+		pitch_speed_weight = 2.0f * _percent_undersped + (1.0f - _percent_undersped) * pitch_speed_weight;
 
-	// Calculate specific energies in units of (m**2/sec**2)
-	_SPE_estimate = _vert_pos_state * CONSTANTS_ONE_G; // potential energy
-	_SKE_estimate = 0.5f * _tas_state * _tas_state; // kinetic energy
+	} else if (!flag.airspeed_enabled) {
+		pitch_speed_weight = 0.0f;
 
-	// Calculate specific energy rates in units of (m**2/sec**3)
-	_SPE_rate = _vert_vel_state * CONSTANTS_ONE_G; // potential energy rate of change
-	_SKE_rate = _tas_state * _tas_rate_filtered;// kinetic energy rate of change
+	}
+
+	// don't allow any weight to be larger than one, as it has the same effect as reducing the control
+	// loop time constant and therefore can lead to a destabilization of that control loop
+	weight.spe_weighting = constrain(2.0f - pitch_speed_weight, 0.f, 1.f);
+	weight.ske_weighting = constrain(pitch_speed_weight, 0.f, 1.f);
+
+	return weight;
 }
 
-void TECS::_update_throttle_setpoint()
+void TECSControl::_updatePitchSetpoint(float dt, const Input &input, const SpecificEnergy &se, Param &param,
+				       const Flag &flag)
 {
-	// Calculate demanded rate of change of total energy, respecting vehicle limits.
-	// We will constrain the value below.
-	_STE_rate_setpoint = _SPE_rate_setpoint + _SKE_rate_setpoint;
+	SpecificEnergyWeighting weight{_updateSpeedAltitudeWeights(param, flag)};
+	/*
+	 * The SKE_weighting variable controls how speed and altitude control are prioritised by the pitch demand calculation.
+	 * A weighting of 1 givea equal speed and altitude priority
+	 * A weighting of 0 gives 100% priority to altitude control and must be used when no airspeed measurement is available.
+	 * A weighting of 2 provides 100% priority to speed control and is used when:
+	 * a) an underspeed condition is detected.
+	 * b) during climbout where a minimum pitch angle has been set to ensure altitude is gained. If the airspeed
+	 * rises above the demanded value, the pitch angle demand is increased by the TECS controller to prevent the vehicle overspeeding.
+	 * The weighting can be adjusted between 0 and 2 depending on speed and altitude accuracy requirements.
+	*/
+	// Calculate the specific energy balance rate demand
+	float seb_rate_setpoint = se.spe.rate_setpoint * weight.spe_weighting - se.ske.rate_setpoint * weight.ske_weighting;
 
-	// Calculate the total energy rate error, applying a first order IIR filter
-	// to reduce the effect of accelerometer noise
-	_STE_rate_error_filter.update(-_SPE_rate - _SKE_rate + _SPE_rate_setpoint + _SKE_rate_setpoint);
-	_STE_rate_error = _STE_rate_error_filter.getState();
+	// Calculate the specific energy balance rate error
+	float seb_rate_error = (se.spe.rate_error * weight.spe_weighting) - (se.ske.rate_error * weight.ske_weighting);
+
+	_debug_output.energy_balance_rate_error = seb_rate_error;
+	_debug_output.energy_balance_rate_sp = seb_rate_setpoint;
+
+	if (param.integrator_gain_pitch > 0.0f) {
+		// Calculate pitch integrator input term
+		float pitch_integ_input = seb_rate_error * param.integrator_gain_pitch;
+
+		// Prevent the integrator changing in a direction that will increase pitch demand saturation
+		if (_pitch_setpoint > param.pitch_max) {
+			pitch_integ_input = min(pitch_integ_input, 0.f);
+
+		} else if (_pitch_setpoint < param.pitch_min) {
+			pitch_integ_input = max(pitch_integ_input, 0.f);
+		}
+
+		// Update the pitch integrator state.
+		_pitch_integ_state = _pitch_integ_state + pitch_integ_input * dt;
+
+	} else {
+		_pitch_integ_state = 0.0f;
+	}
+
+	// Calculate a specific energy correction that doesn't include the integrator contribution
+	float SEB_rate_correction = seb_rate_error * param.pitch_damping_gain + _pitch_integ_state + param.seb_rate_ff *
+				    seb_rate_setpoint;
+
+	// Calculate derivative from change in climb angle to rate of change of specific energy balance
+	const float climb_angle_to_SEB_rate = input.tas * CONSTANTS_ONE_G;
+
+	// During climbout, bias the demanded pitch angle so that a zero speed error produces a pitch angle
+	// demand equal to the minimum pitch angle set by the mission plan. This prevents the integrator
+	// having to catch up before the nose can be raised to reduce excess speed during climbout.
+	if (flag.climbout_mode_active) {
+		SEB_rate_correction += param.pitch_min * climb_angle_to_SEB_rate;
+	}
+
+	// Convert the specific energy balance rate correction to a target pitch angle. This calculation assumes:
+	// a) The climb angle follows pitch angle with a lag that is small enough not to destabilise the control loop.
+	// b) The offset between climb angle and pitch angle (angle of attack) is constant, excluding the effect of
+	// pitch transients due to control action or turbulence.
+	float pitch_setpoint_unc = SEB_rate_correction / climb_angle_to_SEB_rate;
+
+	float pitch_setpoint = constrain(pitch_setpoint_unc, param.pitch_min, param.pitch_max);
+
+	// Comply with the specified vertical acceleration limit by applying a pitch rate limit
+	// NOTE: at zero airspeed, the pitch increment is unbounded
+	const float pitch_increment = dt * param.vert_accel_limit / input.tas;
+	_pitch_setpoint = constrain(pitch_setpoint, _pitch_setpoint - pitch_increment,
+				    _pitch_setpoint + pitch_increment);
+}
+
+void TECSControl::_updateThrottleSetpoint(float dt, const SpecificEnergy &se, const Param &param, const Flag &flag)
+{
+	STELimit limit{_calculateTotalEnergyRateLimit(param)};
+
+	float STE_rate_setpoint = se.spe.rate_setpoint + se.ske.rate_setpoint;
 
 	// Adjust the demanded total energy rate to compensate for induced drag rise in turns.
 	// Assume induced drag scales linearly with normal load factor.
 	// The additional normal load factor is given by (1/cos(bank angle) - 1)
-	_STE_rate_setpoint += _load_factor_correction * (_load_factor - 1.f);
+	STE_rate_setpoint += param.load_factor_correction * (param.load_factor - 1.f);
 
-	_STE_rate_setpoint = constrain(_STE_rate_setpoint, _STE_rate_min, _STE_rate_max);
+	STE_rate_setpoint = constrain(STE_rate_setpoint, limit.STE_rate_min, limit.STE_rate_max);
+
+	_ste_rate = se.spe.rate + se.ske.rate;
+
+	float STE_rate_error_raw = se.spe.rate_error + se.ske.rate_error;
+	// TODO rmeove reset and add correct time intervall
+	_ste_rate_error_filter.setParameters(TECS::DT_DEFAULT, param.ste_rate_time_const);
+	_ste_rate_error_filter.reset(0.0f);
+	_ste_rate_error_filter.update(STE_rate_error_raw);
+	float STE_rate_error{_ste_rate_error_filter.getState()};
+
+	_debug_output.total_energy_rate_error = STE_rate_error;
+	_debug_output.total_energy_rate_sp = STE_rate_setpoint;
 
 	// Calculate a predicted throttle from the demanded rate of change of energy, using the cruise throttle
 	// as the starting point. Assume:
@@ -285,31 +471,33 @@ void TECS::_update_throttle_setpoint()
 	// Specific total energy rate = _STE_rate_min is achieved when throttle is set to _throttle_setpoint_min
 	float throttle_predicted = 0.0f;
 
-	if (_STE_rate_setpoint >= 0) {
+	if (STE_rate_setpoint >= 0) {
 		// throttle is between trim and maximum
-		throttle_predicted = _throttle_trim + _STE_rate_setpoint / _STE_rate_max * (_throttle_setpoint_max - _throttle_trim);
+		throttle_predicted = param.throttle_trim + STE_rate_setpoint / limit.STE_rate_max *
+				     (param.throttle_max - param.throttle_trim);
 
 	} else {
 		// throttle is between trim and minimum
-		throttle_predicted = _throttle_trim + _STE_rate_setpoint / _STE_rate_min * (_throttle_setpoint_min - _throttle_trim);
+		throttle_predicted = param.throttle_trim + STE_rate_setpoint / limit.STE_rate_min *
+				     (param.throttle_min - param.throttle_trim);
 
 	}
 
 	// Calculate gain scaler from specific energy rate error to throttle
-	const float STE_rate_to_throttle = 1.0f / (_STE_rate_max - _STE_rate_min);
+	const float STE_rate_to_throttle = 1.0f / (limit.STE_rate_max - limit.STE_rate_min);
 
 	// Add proportional and derivative control feedback to the predicted throttle and constrain to throttle limits
-	float throttle_setpoint = (_STE_rate_error * _throttle_damping_gain) * STE_rate_to_throttle + throttle_predicted;
-	throttle_setpoint = constrain(throttle_setpoint, _throttle_setpoint_min, _throttle_setpoint_max);
+	float throttle_setpoint = (STE_rate_error * param.throttle_damping_gain) * STE_rate_to_throttle + throttle_predicted;
+	throttle_setpoint = constrain(throttle_setpoint, param.throttle_min, param.throttle_max);
 
 	// Integral handling
-	if (airspeed_sensor_enabled()) {
-		if (_integrator_gain_throttle > 0.0f) {
-			float integ_state_max = _throttle_setpoint_max - throttle_setpoint;
-			float integ_state_min = _throttle_setpoint_min - throttle_setpoint;
+	if (flag.airspeed_enabled) {
+		if (param.integrator_gain_throttle > 0.0f) {
+			float integ_state_max = param.throttle_max - throttle_setpoint;
+			float integ_state_min = param.throttle_min - throttle_setpoint;
 
 			// underspeed conditions zero out integration
-			float throttle_integ_input = (_STE_rate_error * _integrator_gain_throttle) * _dt *
+			float throttle_integ_input = (STE_rate_error * param.integrator_gain_throttle) * dt *
 						     STE_rate_to_throttle * (1.0f - _percent_undersped);
 
 			// only allow integrator propagation into direction which unsaturates throttle
@@ -324,7 +512,7 @@ void TECS::_update_throttle_setpoint()
 			// This will be added to the total throttle demand to compensate for steady state errors
 			_throttle_integ_state = _throttle_integ_state + throttle_integ_input;
 
-			if (_climbout_mode_active) {
+			if (flag.climbout_mode_active) {
 				// During climbout, set the integrator to maximum throttle to prevent transient throttle drop
 				// at end of climbout when we transition to closed loop throttle control
 				_throttle_integ_state = integ_state_max;
@@ -336,7 +524,7 @@ void TECS::_update_throttle_setpoint()
 
 	}
 
-	if (airspeed_sensor_enabled()) {
+	if (flag.airspeed_enabled) {
 		// Add the integrator feedback during closed loop operation with an airspeed sensor
 		throttle_setpoint += _throttle_integ_state;
 
@@ -347,38 +535,81 @@ void TECS::_update_throttle_setpoint()
 	}
 
 	// ramp in max throttle setting with underspeediness value
-	throttle_setpoint = _percent_undersped * _throttle_setpoint_max + (1.0f - _percent_undersped) * throttle_setpoint;
+	throttle_setpoint = _percent_undersped * param.throttle_max + (1.0f - _percent_undersped) * throttle_setpoint;
 
 	// Rate limit the throttle demand
-	if (fabsf(_throttle_slewrate) > 0.01f) {
-		const float throttle_increment_limit = _dt * (_throttle_setpoint_max - _throttle_setpoint_min) * _throttle_slewrate;
-		throttle_setpoint = constrain(throttle_setpoint, _last_throttle_setpoint - throttle_increment_limit,
-					      _last_throttle_setpoint + throttle_increment_limit);
+	if (fabsf(param.throttle_slewrate) > 0.01f) {
+		const float throttle_increment_limit = dt * (param.throttle_max - param.throttle_min) * param.throttle_slewrate;
+		throttle_setpoint = constrain(throttle_setpoint, _throttle_setpoint - throttle_increment_limit,
+					      _throttle_setpoint + throttle_increment_limit);
 	}
 
-	_last_throttle_setpoint = constrain(throttle_setpoint, _throttle_setpoint_min, _throttle_setpoint_max);
+	_throttle_setpoint = constrain(throttle_setpoint, param.throttle_min, param.throttle_max);
 }
 
-void TECS::_detect_uncommanded_descent()
+void TECSControl::resetIntegrals()
+{
+	_pitch_integ_state = 0.0f;
+	_throttle_integ_state = 0.0f;
+}
+
+float TECS::_update_speed_setpoint(const float tas_min, const float tas_max, const float tas_setpoint, const float tas)
+{
+	float new_setpoint{tas_setpoint};
+	float percent_undersped = _control.getPercentUndersped();
+
+	// Set the TAS demand to the minimum value if an underspeed or
+	// or a uncontrolled descent condition exists to maximise climb rate
+	if (_uncommanded_descent_recovery) {
+		new_setpoint = tas_min;
+
+	} else if (percent_undersped > FLT_EPSILON) {
+		// TAS setpoint is reset from external setpoint every time tecs is called, so the interpolation is still
+		// between current setpoint and mininimum airspeed here (it's not feeding the newly adjusted setpoint
+		// from this line back into this method each time).
+		// TODO: WOULD BE GOOD to "functionalize" this library a bit and remove many of these internal states to
+		// avoid the fear of side effects in simple operations like this.
+		new_setpoint = tas_min * percent_undersped + (1.0f - percent_undersped) * tas_setpoint;
+	}
+
+	new_setpoint = constrain(new_setpoint, tas_min, tas_max);
+
+	return new_setpoint;
+}
+
+void TECS::_detect_uncommanded_descent(float throttle_setpoint_max, float altitude, float altitude_setpoint, float tas,
+				       float tas_setpoint)
 {
 	/*
 	 * This function detects a condition that can occur when the demanded airspeed is greater than the
-	 * aircraft can achieve in level flight. When this occurs, the vehicle will continue to reduce height
+	 * aircraft can achieve in level flight. When this occurs, the vehicle will continue to reduce altitude
 	 * while attempting to maintain speed.
 	*/
 
-	// Calculate rate of change of total specific energy
-	const float STE_rate = _SPE_rate + _SKE_rate;
+	// Calculate specific energy demands in units of (m**2/sec**2)
+	float SPE_setpoint = altitude_setpoint * CONSTANTS_ONE_G; // potential energy
+	float SKE_setpoint = 0.5f * altitude_setpoint * altitude_setpoint; // kinetic energy
 
-	const bool underspeed_detected = _percent_undersped > FLT_EPSILON;
+	// Calculate specific energies in units of (m**2/sec**2)
+	float SPE_estimate = altitude * CONSTANTS_ONE_G; // potential energy
+	float SKE_estimate = 0.5f * tas * tas; // kinetic energy
+
+	// Calculate total energy error
+	float SPE_error = SPE_setpoint - SPE_estimate;
+	float SKE_error = SKE_setpoint - SKE_estimate;
+	float STE_error = SPE_error + SKE_error;
+
+
+
+	const bool underspeed_detected = _control.getPercentUndersped() > FLT_EPSILON;
 
 	// If total energy is very low and reducing, throttle is high, and we are not in an underspeed condition, then enter uncommanded descent recovery mode
-	const bool enter_mode = !_uncommanded_descent_recovery && !underspeed_detected && (_STE_error > 200.0f)
-				&& (STE_rate < 0.0f)
-				&& (_last_throttle_setpoint >= _throttle_setpoint_max * 0.9f);
+	const bool enter_mode = !_uncommanded_descent_recovery && !underspeed_detected && (STE_error > 200.0f)
+				&& (_control.getSteRate() < 0.0f)
+				&& (_control.getThrottleSetpoint() >= throttle_setpoint_max * 0.9f);
 
 	// If we enter an underspeed condition or recover the required total energy, then exit uncommanded descent recovery mode
-	const bool exit_mode = _uncommanded_descent_recovery && (underspeed_detected || (_STE_error < 0.0f));
+	const bool exit_mode = _uncommanded_descent_recovery && (underspeed_detected || (STE_error < 0.0f));
 
 	if (enter_mode) {
 		_uncommanded_descent_recovery = true;
@@ -389,243 +620,108 @@ void TECS::_detect_uncommanded_descent()
 	}
 }
 
-void TECS::_update_pitch_setpoint()
+void TECS::initialize(const float altitude, const float altitude_rate, const float equivalent_airspeed)
 {
-	/*
-	 * The SKE_weighting variable controls how speed and height control are prioritised by the pitch demand calculation.
-	 * A weighting of 1 givea equal speed and height priority
-	 * A weighting of 0 gives 100% priority to height control and must be used when no airspeed measurement is available.
-	 * A weighting of 2 provides 100% priority to speed control and is used when:
-	 * a) an underspeed condition is detected.
-	 * b) during climbout where a minimum pitch angle has been set to ensure height is gained. If the airspeed
-	 * rises above the demanded value, the pitch angle demand is increased by the TECS controller to prevent the vehicle overspeeding.
-	 * The weighting can be adjusted between 0 and 2 depending on speed and height accuracy requirements.
-	*/
-
-	// Calculate the specific energy balance rate demand
-	const float SEB_rate_setpoint = _SPE_rate_setpoint * _SPE_weighting - _SKE_rate_setpoint * _SKE_weighting;
-
-	// Calculate the specific energy balance rate error
-	_SEB_rate_error = SEB_rate_setpoint - (_SPE_rate * _SPE_weighting - _SKE_rate * _SKE_weighting);
-
-	// Calculate derivative from change in climb angle to rate of change of specific energy balance
-	const float climb_angle_to_SEB_rate = _tas_state * CONSTANTS_ONE_G;
-
-	if (_integrator_gain_pitch > 0.0f) {
-		// Calculate pitch integrator input term
-		float pitch_integ_input = _SEB_rate_error * _integrator_gain_pitch;
-
-		// Prevent the integrator changing in a direction that will increase pitch demand saturation
-		if (_pitch_setpoint_unc > _pitch_setpoint_max) {
-			pitch_integ_input = min(pitch_integ_input, 0.f);
-
-		} else if (_pitch_setpoint_unc < _pitch_setpoint_min) {
-			pitch_integ_input = max(pitch_integ_input, 0.f);
-		}
-
-		// Update the pitch integrator state.
-		_pitch_integ_state = _pitch_integ_state + pitch_integ_input * _dt;
-
-	} else {
-		_pitch_integ_state = 0.0f;
-	}
-
-	// Calculate a specific energy correction that doesn't include the integrator contribution
-	float SEB_rate_correction = _SEB_rate_error * _pitch_damping_gain + _pitch_integ_state + _SEB_rate_ff *
-				    SEB_rate_setpoint;
-
-	// During climbout, bias the demanded pitch angle so that a zero speed error produces a pitch angle
-	// demand equal to the minimum pitch angle set by the mission plan. This prevents the integrator
-	// having to catch up before the nose can be raised to reduce excess speed during climbout.
-	if (_climbout_mode_active) {
-		SEB_rate_correction += _pitch_setpoint_min * climb_angle_to_SEB_rate;
-	}
-
-	// Convert the specific energy balance rate correction to a target pitch angle. This calculation assumes:
-	// a) The climb angle follows pitch angle with a lag that is small enough not to destabilise the control loop.
-	// b) The offset between climb angle and pitch angle (angle of attack) is constant, excluding the effect of
-	// pitch transients due to control action or turbulence.
-	_pitch_setpoint_unc = SEB_rate_correction / climb_angle_to_SEB_rate;
-
-	float pitch_setpoint = constrain(_pitch_setpoint_unc, _pitch_setpoint_min, _pitch_setpoint_max);
-
-	// Comply with the specified vertical acceleration limit by applying a pitch rate limit
-	// NOTE: at zero airspeed, the pitch increment is unbounded
-	const float pitch_increment = _dt * _vert_accel_limit / _tas_state;
-	_last_pitch_setpoint = constrain(pitch_setpoint, _last_pitch_setpoint - pitch_increment,
-					 _last_pitch_setpoint + pitch_increment);
+	// Init subclasses
+	TECSReferenceModel::AltitudeReferenceState current_state{	.alt = altitude,
+			.alt_rate = altitude_rate};
+	_reference_model.initialize(current_state);
+	_airspeed_filter.initialize(equivalent_airspeed);
+	_control.initialize();
 }
 
-void TECS::_updateTrajectoryGenerationConstraints()
-{
-	_alt_control_traj_generator.setMaxJerk(_jerk_max);
-	_alt_control_traj_generator.setMaxAccel(_vert_accel_limit);
-	_alt_control_traj_generator.setMaxVel(fmax(_max_climb_rate, _max_sink_rate));
-
-	_velocity_control_traj_generator.setMaxJerk(_jerk_max);
-	_velocity_control_traj_generator.setMaxAccelUp(_vert_accel_limit);
-	_velocity_control_traj_generator.setMaxAccelDown(_vert_accel_limit);
-	_velocity_control_traj_generator.setMaxVelUp(_max_sink_rate); // different convention for FW than for MC
-	_velocity_control_traj_generator.setMaxVelDown(_max_climb_rate); // different convention for FW than for MC
-}
-
-void TECS::_calculateHeightRateSetpoint(float altitude_sp_amsl, float height_rate_sp, float target_climbrate,
-					float target_sinkrate, float altitude_amsl)
-{
-	bool control_altitude = true;
-	const bool input_is_height_rate = PX4_ISFINITE(height_rate_sp);
-
-	_velocity_control_traj_generator.setVelSpFeedback(_hgt_rate_setpoint);
-
-	if (input_is_height_rate) {
-		_velocity_control_traj_generator.setCurrentPositionEstimate(altitude_amsl);
-		_velocity_control_traj_generator.update(_dt, height_rate_sp);
-		_hgt_rate_setpoint = _velocity_control_traj_generator.getCurrentVelocity();
-		_smooth_hgt_rate_setpoint = _hgt_rate_setpoint;
-		altitude_sp_amsl = _velocity_control_traj_generator.getCurrentPosition();
-		control_altitude = PX4_ISFINITE(altitude_sp_amsl);
-
-	} else {
-		_velocity_control_traj_generator.reset(0, _hgt_rate_setpoint, _hgt_setpoint);
-	}
-
-
-	if (control_altitude) {
-		runAltitudeControllerSmoothVelocity(altitude_sp_amsl, target_climbrate, target_sinkrate, altitude_amsl);
-
-	} else {
-		_alt_control_traj_generator.setCurrentVelocity(_hgt_rate_setpoint);
-		_alt_control_traj_generator.setCurrentPosition(altitude_amsl);
-		_hgt_setpoint = altitude_amsl;
-	}
-}
-
-void TECS::_initialize_states(float pitch, float throttle_trim, float baro_altitude, float pitch_min_climbout,
-			      float EAS2TAS)
-{
-	if (_pitch_update_timestamp == 0 || _dt > DT_MAX || !_states_initialized) {
-		// On first time through or when not using TECS of if there has been a large time slip,
-		// states must be reset to allow filters to a clean start
-		_vert_vel_state = 0.0f;
-		_vert_pos_state = baro_altitude;
-		_tas_rate_state = 0.0f;
-		_tas_state = _EAS * EAS2TAS;
-		_last_throttle_setpoint = throttle_trim;
-		_last_pitch_setpoint = constrain(pitch, _pitch_setpoint_min, _pitch_setpoint_max);
-		_pitch_setpoint_unc = _last_pitch_setpoint;
-		_TAS_setpoint_last = _EAS * EAS2TAS;
-		_TAS_setpoint_adj = _TAS_setpoint_last;
-		_uncommanded_descent_recovery = false;
-		_STE_rate_error = 0.0f;
-		_hgt_setpoint = baro_altitude;
-
-		resetIntegrals();
-
-		if (_dt > DT_MAX || _dt < DT_MIN) {
-			_dt = DT_DEFAULT;
-		}
-
-		resetTrajectoryGenerators(baro_altitude);
-
-	} else if (_climbout_mode_active) {
-		// During climbout use the lower pitch angle limit specified by the
-		// calling controller
-		_pitch_setpoint_min	   = pitch_min_climbout;
-
-		// throttle lower limit is set to a value that prevents throttle reduction
-		_throttle_setpoint_min  = _throttle_setpoint_max - 0.01f;
-
-		// airspeed demand states are set to track the measured airspeed
-		_TAS_setpoint_last      = _EAS * EAS2TAS;
-		_TAS_setpoint_adj       = _EAS * EAS2TAS;
-
-		_hgt_setpoint = baro_altitude;
-
-		_uncommanded_descent_recovery = false;
-	}
-
-	// filter specific energy rate error using first order filter with 0.5 second time constant
-	_STE_rate_error_filter.setParameters(DT_DEFAULT, _STE_rate_time_const);
-	_STE_rate_error_filter.reset(0.0f);
-
-	// filter true airspeed rate using first order filter with 0.5 second time constant
-	_TAS_rate_filter.setParameters(DT_DEFAULT, _speed_derivative_time_const);
-	_TAS_rate_filter.reset(0.0f);
-
-	_states_initialized = true;
-}
-
-void TECS::_update_STE_rate_lim()
-{
-	// Calculate the specific total energy upper rate limits from the max throttle climb rate
-	_STE_rate_max = math::max(_max_climb_rate, FLT_EPSILON) * CONSTANTS_ONE_G;
-
-	// Calculate the specific total energy lower rate limits from the min throttle sink rate
-	_STE_rate_min = - math::max(_min_sink_rate, FLT_EPSILON) * CONSTANTS_ONE_G;
-}
-
-void TECS::update_pitch_throttle(float pitch, float baro_altitude, float hgt_setpoint,
-				 float EAS_setpoint, float equivalent_airspeed, float eas_to_tas, bool climb_out_setpoint, float pitch_min_climbout,
-				 float throttle_min, float throttle_max, float throttle_trim,
-				 float pitch_limit_min, float pitch_limit_max,
-				 float target_climbrate, float target_sinkrate, float hgt_rate_sp)
+void TECS::update(float pitch, float altitude, float hgt_setpoint, float EAS_setpoint, float equivalent_airspeed,
+		  float eas_to_tas, bool climb_out_setpoint, float pitch_min_climbout, float throttle_min, float throttle_setpoint_max,
+		  float throttle_trim, float pitch_limit_min, float pitch_limit_max, float target_climbrate, float target_sinkrate,
+		  const float speed_deriv_forward, float hgt_rate, float hgt_rate_sp)
 {
 	// Calculate the time since last update (seconds)
 	uint64_t now = hrt_absolute_time();
-	_dt = fmaxf((now - _pitch_update_timestamp) * 1e-6f, DT_MIN);
+	float dt = (now - _update_timestamp) * 1e-6f;
 
-	// Set class variables from inputs
-	_throttle_setpoint_max = throttle_max;
-	_throttle_setpoint_min = throttle_min;
-	_pitch_setpoint_max = pitch_limit_max;
-	_pitch_setpoint_min = pitch_limit_min;
-	_climbout_mode_active = climb_out_setpoint;
-	_throttle_trim = throttle_trim;
+	if (dt < DT_MIN) {
+		// Update intervall too small, do not update. Assume constant states/output in this case.
+		return;
+	}
 
-	// Initialize selected states and variables as required
-	_initialize_states(pitch, throttle_trim, baro_altitude, pitch_min_climbout, eas_to_tas);
+	if (dt > DT_MAX || _update_timestamp == 0UL) {
+		// Update time intervall too large, can't guarantee sanity of state updates anymore. reset the control loop.
+		initialize(altitude, hgt_rate, equivalent_airspeed);
 
-	_updateTrajectoryGenerationConstraints();
+	} else {
+		// Update airspeedfilter submodule
+		TECSAirspeedFilter::Input airspeed_input{ .equivalent_airspeed = equivalent_airspeed,
+				.equivalent_airspeed_rate = speed_deriv_forward / eas_to_tas};
+		_airspeed_param.equivalent_airspeed_trim = _equivalent_airspeed_trim;
+		_airspeed_filter.update(dt, airspeed_input, _airspeed_param, _airspeed_enabled);
+		TECSAirspeedFilter::AirspeedFilterState eas = _airspeed_filter.getState();
 
-	// Update the true airspeed state estimate
-	_update_speed_states(EAS_setpoint, equivalent_airspeed, eas_to_tas);
+		// Update Reference model submodule
+		TECSReferenceModel::AltitudeReferenceState setpoint{ .alt = hgt_setpoint,
+				.alt_rate = hgt_rate_sp
+								   };
+		_reference_param.target_climbrate = target_climbrate;
+		_reference_param.target_sinkrate = target_sinkrate;
+		_reference_model.update(dt, setpoint, altitude, _reference_param);
+		TECSControl::Setpoint control_setpoint;
+		control_setpoint.altitude_reference = _reference_model.getAltitudeReference();
+		control_setpoint.altitude_rate_setpoint = _reference_model.getAltitudeRateReference();
 
-	// Calculate rate limits for specific total energy
-	_update_STE_rate_lim();
+		// Calculate the demanded true airspeed
+		// TODO this function should not be in the module. Only give feedback that the airspeed can't be achieved.
+		control_setpoint.tas_setpoint = _update_speed_setpoint(eas_to_tas * _equivalent_airspeed_min,
+						eas_to_tas * _equivalent_airspeed_max, EAS_setpoint * eas_to_tas, eas_to_tas * eas.speed);
 
-	// Detect an underspeed condition
-	_detect_underspeed();
+		TECSControl::Input control_input{ .altitude = altitude,
+						  .altitude_rate = hgt_rate,
+						  .tas = eas_to_tas * eas.speed,
+						  .tas_rate = eas_to_tas * eas.speed_rate
+						};
+		_control_param.equivalent_airspeed_trim = _equivalent_airspeed_trim;
+		_control_param.tas_min = eas_to_tas * _equivalent_airspeed_min;
+		_control_param.pitch_max = pitch_limit_max;
+		_control_param.pitch_min = pitch_limit_min;
+		_control_param.throttle_trim = throttle_trim;
+		_control_param.throttle_max = throttle_setpoint_max;
+		_control_param.throttle_min = throttle_min;
+		TECSControl::Flag control_flag{ .airspeed_enabled = _airspeed_enabled,
+						.climbout_mode_active = climb_out_setpoint,
+						.detect_underspeed_enabled = _detect_underspeed_enabled
+					      };
+		_control.update(dt, control_setpoint, control_input, _control_param, control_flag);
 
-	_update_speed_height_weights();
+		// Detect an uncommanded descent caused by an unachievable airspeed demand
+		_detect_uncommanded_descent(throttle_setpoint_max, altitude, hgt_setpoint, equivalent_airspeed * eas_to_tas,
+					    control_setpoint.tas_setpoint);
 
-	// Detect an uncommanded descent caused by an unachievable airspeed demand
-	_detect_uncommanded_descent();
+		TECSControl::DebugOutput control_status = _control.getDebugOutput();
+		_debug_status.altitude_rate_control = control_status.altitude_rate_control;
+		_debug_status.energy_balance_rate_error = control_status.energy_balance_rate_error;
+		_debug_status.energy_balance_rate_sp = control_status.energy_balance_rate_sp;
+		_debug_status.total_energy_rate_error = control_status.total_energy_rate_error;
+		_debug_status.total_energy_rate_sp = control_status.total_energy_rate_sp;
+		_debug_status.true_airspeed_derivative_control = control_status.true_airspeed_derivative_control;
+		_debug_status.true_airspeed_filtered = eas_to_tas * eas.speed;
+		_debug_status.true_airspeed_derivative = eas_to_tas * eas.speed_rate;
+		_debug_status.altitude_sp = control_setpoint.altitude_reference.alt;
+		_debug_status.altitude_rate = control_setpoint.altitude_reference.alt_rate;
+		_debug_status.altitude_rate_setpoint = control_setpoint.altitude_rate_setpoint;
+	}
 
-	// Calculate the demanded true airspeed
-	_update_speed_setpoint();
 
-	_calculateHeightRateSetpoint(hgt_setpoint, hgt_rate_sp, target_climbrate, target_sinkrate, baro_altitude);
-
-	// Calculate the specific energy values required by the control loop
-	_update_energy_estimates();
-
-	// Calculate the throttle demand
-	_update_throttle_setpoint();
-
-	// Calculate the pitch demand
-	_update_pitch_setpoint();
 
 	// Update time stamps
-	_pitch_update_timestamp = now;
+	_update_timestamp = now;
+
 
 	// Set TECS mode for next frame
-	if (_percent_undersped > FLT_EPSILON) {
+	if (_control.getPercentUndersped() > FLT_EPSILON) {
 		_tecs_mode = ECL_TECS_MODE_UNDERSPEED;
 
 	} else if (_uncommanded_descent_recovery) {
 		_tecs_mode = ECL_TECS_MODE_BAD_DESCENT;
 
-	} else if (_climbout_mode_active) {
+	} else if (climb_out_setpoint) {
 		_tecs_mode = ECL_TECS_MODE_CLIMBOUT;
 
 	} else {
@@ -633,26 +729,6 @@ void TECS::update_pitch_throttle(float pitch, float baro_altitude, float hgt_set
 		_tecs_mode = ECL_TECS_MODE_NORMAL;
 	}
 
+	_debug_status.tecs_mode = _tecs_mode;
 }
 
-void TECS::_update_speed_height_weights()
-{
-	// Calculate the weight applied to control of specific kinetic energy error
-	float pitch_speed_weight = constrain(_pitch_speed_weight, 0.0f, 2.0f);
-
-	if (_climbout_mode_active && airspeed_sensor_enabled()) {
-		pitch_speed_weight = 2.0f;
-
-	} else if (_percent_undersped > FLT_EPSILON && airspeed_sensor_enabled()) {
-		pitch_speed_weight = 2.0f * _percent_undersped + (1.0f - _percent_undersped) * pitch_speed_weight;
-
-	} else if (!airspeed_sensor_enabled()) {
-		pitch_speed_weight = 0.0f;
-
-	}
-
-	// don't allow any weight to be larger than one, as it has the same effect as reducing the control
-	// loop time constant and therefore can lead to a destabilization of that control loop
-	_SPE_weighting = constrain(2.0f - pitch_speed_weight, 0.f, 1.f);
-	_SKE_weighting = constrain(pitch_speed_weight, 0.f, 1.f);
-}
